# packages/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/posts/channels": {
      "get": {
        "description": "Channels are fully dynamic. A channel added in the CMS appears here — and therefore in the app — without any code change.",
        "operationId": "PostsController_channels_v1",
        "parameters": [
          {
            "name": "locale",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "de",
                "en"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Active channels ordered by sortOrder, then name.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelsResponseDto"
                }
              }
            }
          }
        },
        "summary": "List active channels.",
        "tags": [
          "posts"
        ]
      }
    },
    "/v1/posts/tags": {
      "get": {
        "description": "Tags are fully dynamic. A tag added in the CMS appears here — and therefore in the app — without any code change.",
        "operationId": "PostsController_tags_v1",
        "parameters": [
          {
            "name": "locale",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "de",
                "en"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Active tags ordered by sortOrder, then name.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagsResponseDto"
                }
              }
            }
          }
        },
        "summary": "List active tags.",
        "tags": [
          "posts"
        ]
      }
    },
    "/v1/posts/events": {
      "get": {
        "description": "Filter for posts with tag \"event\". Accepts optional `from` and `to` date bounds and channel filters. Interval overlap: eventStart <= to AND (eventEnd ?? eventStart) >= from. Sorted deterministically by eventStart asc, slug asc.",
        "operationId": "PostsController_events_v1",
        "parameters": [
          {
            "name": "locale",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "de",
                "en"
              ],
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Max 50.",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "channels",
            "required": false,
            "in": "query",
            "description": "Comma-separated channel slugs. At most 25, each at most 100 characters.",
            "schema": {}
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "YYYY-MM-DD",
            "schema": {}
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "YYYY-MM-DD",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostListResponseDto"
                }
              }
            }
          }
        },
        "summary": "List event posts within a time range.",
        "tags": [
          "posts"
        ]
      }
    },
    "/v1/posts/{slug}": {
      "get": {
        "description": "Unknown content block types are removed server-side and reported in meta.droppedBlockTypes, so a new CMS block type can never break the detail screen.",
        "operationId": "PostsController_detail_v1",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "de",
                "en"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDetailResponseDto"
                }
              }
            }
          }
        },
        "summary": "Fetch one post by its stable slug.",
        "tags": [
          "posts"
        ]
      }
    },
    "/v1/posts": {
      "get": {
        "description": "An absent `channels` parameter means all active channels; a present but empty `channels=` means the user deselected everything and yields an empty list. An absent `tags` parameter applies no tag filter; a present but empty `tags=` yields an empty list. Both filters combine with AND. Each entry carries its sanitised `content`, so a feed can render posts inline without a request per card; unknown block types are removed server-side and reported once in `meta.droppedBlockTypes`.",
        "operationId": "PostsController_list_v1",
        "parameters": [
          {
            "name": "locale",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "de",
                "en"
              ],
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Max 50.",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "tags",
            "required": false,
            "in": "query",
            "description": "Comma-separated tag slugs. At most 25, each at most 100 characters.",
            "schema": {}
          },
          {
            "name": "channels",
            "required": false,
            "in": "query",
            "description": "Comma-separated channel slugs. At most 25, each at most 100 characters.",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostListResponseDto"
                }
              }
            }
          }
        },
        "summary": "List published posts.",
        "tags": [
          "posts"
        ]
      }
    },
...
  },
  "info": {
    "title": "Campus Köthen API",
    "description": "Public read-only API of the Campus Köthen app.\n\nCampus Köthen is an independent, unofficial campus app. It is neither developed\nnor operated by Hochschule Anhalt, nor is it officially endorsed by the university.\n\nAll content endpoints accept `locale=de|en` (default `de`) and report\n`requestedLocale`, `resolvedLocale` and `translationFallback` in their metadata.\nCanteen dish text comes from a German-only source and is never machine-translated.",
    "version": "0.1.0",
    "contact": {
      "name": "Campus Köthen App",
      "url": "https://github.com/Leviora-Studio/campus-koethen",
      "email": ""
    },
    "license": {
      "name": "AGPL-3.0-only",
      "url": "https://www.gnu.org/licenses/agpl-3.0.html"
    }
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "ChannelDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "example": "campus-news",
            "description": "Stable, never localised."
          },
          "name": {
            "type": "string",
            "example": "Campus News"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "colorHex": {
            "type": "string",
            "example": "#5B3FD0",
            "pattern": "^#[0-9A-Fa-f]{6}$"
          },
          "sortOrder": {
            "type": "number",
            "example": 10
          },
          "defaultSubscribed": {
            "type": "boolean",
            "description": "Applied by the client exactly once, when the slug first appears."
          },
          "publicCalendarSlug": {
            "type": "string",
            "nullable": true,
            "description": "Slug of the associated public Google calendar, if linked 1-to-1."
          }
        },
        "required": [
          "slug",
          "name",
          "colorHex",
          "sortOrder",
          "defaultSubscribed"
        ]
      },
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "example": 20
          },
          "total": {
            "type": "number",
            "example": 42
          },
          "totalPages": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "page",
          "pageSize",
          "total",
          "totalPages"
        ]
      },
      "ResponseMetaDto": {
        "type": "object",
        "properties": {
          "requestedLocale": {
            "type": "string",
            "enum": [
              "de",
              "en"
            ],
            "description": "Locale the client asked for."
          },
          "resolvedLocale": {
            "type": "string",
            "enum": [
              "de",
              "en"
            ],
            "description": "Locale actually served."
          },
          "translationFallback": {
            "type": "boolean",
            "description": "True when at least one delivered field came from the German fallback although another locale was requested."
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          },
          "droppedBlockTypes": {
            "description": "Content block types removed because the MVP client does not render them.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "lastSuccessfulSyncAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Last successful canteen synchronisation. Null means: never succeeded."
          },
          "dataStale": {
            "type": "boolean",
            "description": "True when the data is older than the configured threshold."
          },
          "from": {
            "type": "string",
            "description": "Start of the resolved date range (YYYY-MM-DD)."
          },
          "to": {
            "type": "string",
            "description": "End of the resolved date range (YYYY-MM-DD)."
          },
          "truncated": {
            "type": "boolean",
            "description": "True when the server hit its own result ceiling and deliberately cut the list short. The client must say so rather than present the result as complete."
          },
          "featureEnabled": {
            "type": "boolean",
            "description": "False when the timetable integration is switched off server-side."
          },
          "dataState": {
            "type": "string",
            "enum": [
              "ready",
              "pending",
              "unavailable"
            ],
            "description": "ready = data present; pending = enabled but not yet synchronised; unavailable = disabled or permanently without data."
          },
          "timezone": {
            "type": "string",
            "description": "Business timezone the wall-clock times belong to.",
            "example": "Europe/Berlin"
          }
        },
        "required": [
          "requestedLocale",
          "resolvedLocale",
          "translationFallback"
        ]
      },
      "ChannelsResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChannelDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "TagDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "example": "event",
            "description": "Stable, never localised."
          },
          "name": {
            "type": "string",
            "example": "Event"
          }
        },
        "required": [
          "slug",
          "name"
        ]
      },
      "TagsResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "ImageDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "example": "https://cdn.example/hero.jpg",
            "description": "Always https."
          },
          "alternativeText": {
            "type": "string",
            "nullable": true
          },
          "width": {
            "type": "number",
            "nullable": true
          },
          "height": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "url",
          "alternativeText",
          "width",
          "height"
        ]
      },
      "TagRefDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "slug",
          "name"
        ]
      },
      "ChannelRefDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "colorHex": {
            "type": "string"
          }
        },
        "required": [
          "slug",
          "name",
          "colorHex"
        ]
      },
      "PostListItemDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "example": "semesterstart-2026"
          },
          "title": {
            "type": "string"
          },
          "publishedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "heroImage": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ImageDto"
              }
            ]
          },
          "tag": {
            "description": "Mandatory single tag classification.",
            "allOf": [
              {
                "$ref": "#/components/schemas/TagRefDto"
              }
            ]
          },
          "primaryChannel": {
            "description": "Mandatory primary publishing channel.",
            "allOf": [
              {
                "$ref": "#/components/schemas/ChannelRefDto"
              }
            ]
          },
          "channels": {
            "description": "All channels this post is published to.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChannelRefDto"
            }
          },
          "sourceName": {
            "type": "string",
            "nullable": true
          },
          "sourceUrl": {
            "type": "string",
            "nullable": true,
            "description": "Validated https URL of the original source, or null."
          },
          "content": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            },
            "description": "Sanitised content blocks, delivered with the LIST entry so a feed can render the post inline without a request per card. Only paragraph, heading, list, quote and image survive; anything else is removed server-side and reported in meta.droppedBlockTypes."
          },
          "eventStart": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "eventEnd": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "eventAllDay": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "slug",
          "title",
          "tag",
          "primaryChannel",
          "channels",
          "content",
          "eventAllDay"
        ]
      },
      "PostListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostListItemDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "PostDetailDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "example": "semesterstart-2026"
          },
          "title": {
            "type": "string"
          },
          "publishedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "heroImage": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ImageDto"
              }
            ]
          },
          "tag": {
            "description": "Mandatory single tag classification.",
            "allOf": [
              {
                "$ref": "#/components/schemas/TagRefDto"
              }
            ]
          },
          "primaryChannel": {
            "description": "Mandatory primary publishing channel.",
            "allOf": [
              {
                "$ref": "#/components/schemas/ChannelRefDto"
              }
            ]
          },
          "channels": {
            "description": "All channels this post is published to.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChannelRefDto"
            }
          },
          "sourceName": {
            "type": "string",
            "nullable": true
          },
          "sourceUrl": {
            "type": "string",
            "nullable": true,
            "description": "Validated https URL of the original source, or null."
          },
          "content": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            },
            "description": "Sanitised content blocks, delivered with the LIST entry so a feed can render the post inline without a request per card. Only paragraph, heading, list, quote and image survive; anything else is removed server-side and reported in meta.droppedBlockTypes."
          },
          "eventStart": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "eventEnd": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "eventAllDay": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "slug",
          "title",
          "tag",
          "primaryChannel",
          "channels",
          "content",
          "eventAllDay"
        ]
      },
      "PostDetailResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PostDetailDto"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
    }
  }
}